Fix pass down to correct component
