+++ src/components/AddTikTokForm/AddTikTokForm.js
@@ -1,16 +1,10 @@
 import { useState } from 'react';
 import { Dialog, DialogTitle, DialogContent, DialogContentText, FormControl, TextField, DialogActions, Button } from "@material-ui/core";
-import { Autocomplete } from '@material-ui/lab';
 import PropTypes from 'prop-types';
 
-const AddTikTokForm = ({ status, addTikTok, closeForm, retrieveAuthors }) => {
+const AddTikTokForm = ({ status, addTikTok, closeForm }) => {
   const [tikTokInput, setTikTokInput] = useState('');
 
-  const writeDataListOptions = () => {
-    return retrieveAuthors().map(item => {
-      return <option value={item}/>
-    })
-  }
 
   return (
     <Dialog open={ status }>
@@ -29,12 +23,9 @@ const AddTikTokForm = ({ status, addTikTok, closeForm, retrieveAuthors }) => {
               event => setTikTokInput(event.target.value)
             }
             id="inputTikTok"
-            inputProps={{list: "authors"}}
             autoFocus
           />
-          <datalist id="authors">
-            { writeDataListOptions() }
-          </datalist>
+          
         </FormControl>
         <DialogContentText>
           Paste any tiktok in the input above. No worries about formatting, we can handle it for you!
